Handle and log invalid IP address errors
If the filter get any invalid input for the anonymize_ipv4_network
mode, it will crash the whole Logstash pipeline because of an
unhandled ArgumentError thrown by the IPAddr class.
Furthermore, the causing event got lost during the crash, so
post-mortem analysis were not possible.

Now, ArgumentError is catched and an error message is logged.
The value to be anonymized is kept in the event in the identified form
and the event gets processed further as usual.

Fixes #3.